kingpin-callofduty-functions/src/requirements.txt
callofduty.py==1.2.1
firebase-admin===4.3.0
google-cloud-firestore==1.8.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.29
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
numpy==1.19.1
python-dotenv==0.14.0
rfc3986==1.4.0
scipy==1.5.2
slackclient==2.9.1
sniffio==1.1.0
